+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Version - 0]
 
+## Improve [v0.10.3] - Release Date: 2023-12-18
+
+- added search params in order mangememt menu
+
+## Improve [v0.10.2] - Release Date: 2023-12-18
+
+- set interval for order list
+
 ## Resolved [v0.10.1] - Release Date: 2023-12-13
 
 - fixed highlight news view count